requirements.txt
requests~=2.32.3
dotenv~=0.9.9
python-dotenv~=1.1.0
pandas~=2.3.1
psycopg2-binary~=2.9.10